is-enabled: adding passthrough providers
So here's the deal: certain providers will double count because
of the way passthru calls use segment.js. In order to avoid this,
we *disable* ones which have no client-side equivalent.

This commit disables calls to outbound, librato, and the webhooks
integrations. Any of these will make a second call to segment.js
with the integration explicitly enabled.