Change Facade.library() logic
As currently written, Facade.library() will crash when called from Facade.device() if options.library is an object without a name property. This is because we will pass over both of the if statements in Facade.library() and attempt to run .indexOf() inside of Facade.device() on either an object literal or undefined.

This adds validation to ensure that library.name exists and is a string.